Import Makefile from Kubernetes Master, Adds release target
`check-path` will sniff the ENV for $KUBERNETES_MASTER_BZR envvar. If it
is not found, it will not allow the release process to continue.

`make release` will kick off the release process of installing the
git-vendor python module, and passing off the $KUBERNETES_MASTER_BZR env
location to git-vendor. *note* the constraints of git-vendor still
apply. The repository must have a tag to export and cannot be in a dirty
state.

This is CI-capable so long as CI doesn't alter any files in the
repository prior to "blessing" a release. At which point, we can
automate the release process by setting the ENVVAR for the job.