Removed the overlay_type from relationship validation because this key was not being used in an obvious way, and terminated the relationships usefulness. Leaving the kubernetes minions in an inconsistent state, without the ability to start up.
This leads into a future modification request for an action to spin up
the minion service, and kick off the register command regardless of the
success of the relationship hook. The state of the service pool should
never be dependent on a relationship hook that doesn't get obviously
re-run.

Proposed:   juju do minion-start (this will kick off the appropriate
relationship hook to perform the registration)